Move saveChanges for agendaitem/subcase to service
