+++ app/components/agenda/agendaitem/agendaitem-case-panel/agendaitem-case-panel-edit.js
@@ -1,7 +1,6 @@
 import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
-import { saveChanges as saveSubcaseTitles } from 'frontend-kaleidos/utils/agendaitem-utils';
 import { trimText } from 'frontend-kaleidos/utils/trim-util';
 import { task } from 'ember-concurrency';
 
@@ -15,6 +14,7 @@ import { task } from 'ember-concurrency';
 export default class AgendaitemCasePanelEdit extends Component {
   @service store;
   @service pieceAccessLevelService;
+  @service saveAgendaitemChanges;
 
   propertiesToSet = Object.freeze(['title', 'shortTitle', 'comment']);
 
@@ -54,12 +54,11 @@ export default class AgendaitemCasePanelEdit extends Component {
       confidential: this.args.subcase?.confidential,
     };
 
-    yield saveSubcaseTitles(
+    yield this.saveAgendaitemChanges.saveChanges(
       this.args.agendaitem,
       propertiesToSetOnAgendaitem,
       propertiesToSetOnSubcase,
       shouldResetFormallyOk,
-      this.store,
     );
     if (this.args.subcase && this.args.subcase.confidential) {
       yield this.pieceAccessLevelService.updateDecisionsAccessLevelOfSubcase(this.args.subcase);
