+++ app/components/subcase/subcase-titles-panel/subcase-titles-panel-edit.js
@@ -1,7 +1,6 @@
 import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
-import { saveChanges as saveSubcaseTitles } from 'frontend-kaleidos/utils/agendaitem-utils';
 import { trimText } from 'frontend-kaleidos/utils/trim-util';
 import { task } from 'ember-concurrency';
 
@@ -13,6 +12,7 @@ import { task } from 'ember-concurrency';
 export default class SubcaseTitlesPanelEdit extends Component {
   @service store;
   @service pieceAccessLevelService;
+  @service saveAgendaitemChanges;
 
   propertiesToSet = Object.freeze(['title', 'shortTitle', 'confidential']);
 
@@ -40,12 +40,11 @@ export default class SubcaseTitlesPanelEdit extends Component {
       confidential: this.args.subcase.confidential,
     };
 
-    yield saveSubcaseTitles(
+    yield this.saveAgendaitemChanges.saveChanges(
       this.args.subcase,
       propertiesToSetOnAgendaitem,
       propertiesToSetOnSubcase,
       true,
-      this.store,
     );
     if (this.args.subcase.confidential) {
       yield this.pieceAccessLevelService.updateDecisionsAccessLevelOfSubcase(this.args.subcase);
